# src/main/resources/static/boat4you_ws_entities_crud.openapi.yaml
openapi: 3.0.3
info:
  title: Boat4You Web Service
  description: |-
    API definition for Boat4You Web Service. \
    Contains operations related to Users
  version: 0.0.1
servers:
  - url: https://localhost:8443
    description: Local deployment
...
paths:
...
  /yachts/{id}:
    get:
      tags:
        - Yachts
      summary: Get yacht by yacht id
      description: Get yacht by yacht id
      operationId: getYacht
      parameters:
        - name: id
          in: path
          description: Id of the yacht to be fetched
          required: true
          schema:
            type: integer
            format: int64
        - name: dateFrom
          in: query
          description: Searches price from this date forward (today is a default value)
          required: false
          schema:
            type: string
            format: date
            example: 2025-04-25
        - name: dateTo
          in: query
          description: Searches price up to this date forward
          required: false
          schema:
            type: string
            format: date
            example: 2025-04-25
        - name: currency
          in: query
          description: Currency to convert price into
          required: false
          schema:
            type: string
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/YachtDetailsDto'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: 'boat4you_ws_common.openapi.yaml#/components/schemas/ErrorSchema'
        '404':
          description: Yacht not found
        '500':
          description: Internal server error
  /yachts:
    get:
      tags:
        - Yachts
      summary: Fetch all yachts by search criteria
      description: Fetch all yachts by search criteria
      operationId: getYachts
      parameters:
        - in: query
          name: locationId
          schema:
            type: string
          required: false
          example: l-15
          description: Id of the wanted locationId. First letter always signals if the location is country (c), region (r) or marina (l)
        - in: query
          name: modelId
          schema:
            type: integer
          required: false
        - in: query
          name: charterType
          schema:
            type: string
          required: false
          example: Crewed
        - in: query
          name: vesselType
          schema:
            $ref: '#/components/schemas/VesselType'
          required: false
          example: CATAMARAN
        - in: query
          name: manufacturer
          required: false
          schema:
            type: array
            items:
              type: integer
              format: int64
        - in: query
          name: model
          required: false
          schema:
            type: array
            items:
              type: integer
              format: int64
        - in: query
          name: mainSailType
          schema:
            type: array
            items:
              $ref: '#/components/schemas/SailTypeEnum'
          required: false
          example: CLASSIC_SAIL
        - in: query
          name: minBuildYear
          schema:
            type: integer
          required: false
          example: 2015
        - in: query
          name: maxBuildYear
          schema:
            type: integer
          required: false
          example: 2022
        - in: query
          name: minPersons
          schema:
            type: integer
          required: false
          example: 4
        - in: query
          name: maxPersons
          schema:
            type: integer
          required: false
          example: 6
        - in: query
          name: minCabins
          schema:
            type: integer
          required: false
          example: 6
        - in: query
          name: maxCabins
          schema:
            type: integer
          required: false
          example: 6
        - in: query
          name: minBerths
          schema:
            type: integer
          required: false
          example: 6
        - in: query
          name: maxBerths
          schema:
            type: integer
          required: false
          example: 6
        - in: query
          name: minLength
          schema:
            type: integer
          required: false
          example: 6
        - in: query
          name: maxLength
          schema:
            type: integer
          required: false
          example: 6
        - in: query
          name: minPrice
          schema:
            type: integer
          required: false
          example: 6
        - in: query
          name: maxPrice
          schema:
            type: integer
          required: false
          example: 6
        - in: query
          name: startDate
          schema:
            type: string
            format: date
            example: 2025-04-28
          required: false
        - in: query
          name: endDate
          schema:
            type: string
            format: date
            example: 2025-04-28
          required: false
        - in: query
          name: minWc
          schema:
            type: integer
          required: false
          example: 6
        - in: query
          name: maxWc
          schema:
            type: integer
          required: false
          example: 6
        - in: query
          name: minEnginePower
          schema:
            type: integer
          required: false
          example: 6
        - in: query
          name: maxEnginePower
          schema:
            type: integer
          required: false
          example: 6
        - in: query
          name: currency
          schema:
            type: string
          required: false
          example: USD
        - in: query
          name: sortBy
          schema:
            type: string
          required: false
        - in: query
          name: sortDirection
          schema:
            type: string
          required: false
        - in: query
          name: page
          schema:
            type: integer
            default: 0
          required: false
        - in: query
          name: size
          schema:
            type: integer
            default: 9
          required: false
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: object
                description: Fetch all yachts by search criteria result
                required:
                  - content
                  - page
                properties:
                  content:
                    description: Yachts list
                    type: array
                    items:
                      $ref: '#/components/schemas/YachtSearchResponseDto'
                  page:
                    $ref: 'boat4you_ws_common.openapi.yaml#/components/schemas/InternalPageModel'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: 'boat4you_ws_common.openapi.yaml#/components/schemas/ErrorSchema'
        '403':
          description: Not authorized
        '500':
          description: Internal server error
...
  /yachts/{id}/availability:
    get:
      tags:
        - Yachts
      summary: Fetches offers for a specific yacht
      description: Fetches offers for a specific yacht
      operationId: getYachtAvailability
      parameters:
        - name: id
          in: path
          description: Id of the yacht
          required: true
          schema:
            type: integer
            format: int64
        - name: month
          in: query
          required: false
          schema:
            type: integer
            format: date
            example: 2025-04-25
        - name: year
          in: query
          required: true
          schema:
            type: integer
            format: date
            example: 2025-04-25
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: object
                description: Fetches offers for a specific yacht
                required:
                  - content
                  - totalElements
                properties:
                  content:
                    description: Offer list
                    type: array
                    items:
                      $ref: '#/components/schemas/OfferDto'
                  totalElements:
                    description: Total number of entities
                    type: integer
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: 'boat4you_ws_common.openapi.yaml#/components/schemas/ErrorSchema'
        '404':
          description: Yacht not found
        '500':
          description: Internal server error
  /yachts/{yachtId}/offers:
    get:
      tags:
        - Yachts
      summary: Fetches offers for a specific yacht
      description: Fetches offers for a specific yacht
      operationId: getYachtOffers
      parameters:
        - name: yachtId
          in: path
          description: Id of the yacht
          required: true
          schema:
            type: integer
            format: int64
        - name: dateFrom
          in: query
          description: Searches price from this date forward (today is a default value)
          required: false
          schema:
            type: string
            format: date
            example: 2025-04-25
        - name: dateTo
          in: query
          description: Searches price up to this date forward
          required: false
          schema:
            type: string
            format: date
            example: 2025-04-25
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: object
                description: Fetches offers for a specific yacht
                required:
                  - content
                  - totalElements
                properties:
                  content:
                    description: Offer list
                    type: array
                    items:
                      $ref: '#/components/schemas/OfferDto'
                  totalElements:
                    description: Total number of entities
                    type: integer
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: 'boat4you_ws_common.openapi.yaml#/components/schemas/ErrorSchema'
        '404':
          description: Yacht not found
        '500':
          description: Internal server error
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
    SailTypeEnum:
      type: integer
      enum:
        - UNKNOWN
        - CLASSIC_SAIL
        - ROLLING_SAIL
...
    VesselType:
      type: integer
      enum:
        - OTHER
        - CATAMARAN
        - GULET
        - HOUSE_BOAT
        - LUXURY_CATAMARAN
        - LUXURY_MOTOR_YACHT
        - LUXURY_SAILING_YACHT
        - MINI_CRUISER
        - MOTORBOAT
        - MOTOR_YACHT
        - MOTORSAILER
        - POWER_CATAMARAN
        - SAILING_YACHT
        - TRIMARAN
...
    OfferStatus:
      type: integer
      enum:
        - UNKNOWN
        - FREE
        - OPTION
        - UNAVAILABLE
...
    YachtDetailsDto:
      type: object
      required:
        - name
        - model
        - agency
        - location
        - yachtImages
      properties:
        id:
          type: integer
          format: int64
          example: 10
        name:
          type: string
          example: Big Blue Yacht
        buildYear:
          type: integer
          example: 2011
        maxPersons:
          type: integer
          example: 6
        cabins:
          type: integer
          example: 4
        wc:
          type: integer
          example: 2
        berths:
          type: integer
          example: 4
        enginePower:
          type: integer
          example: 100
        fuelTank:
          type: integer
          example: 80
        waterTank:
          type: integer
          example: 120
        beam:
          type: integer
          example: 120
        mainSailType:
          type: integer
          $ref: '#/components/schemas/SailTypeEnum'
        length:
          type: integer
          example: 15
        model:
          type: string
        clientPriceEur:
          type: integer
          example: 100
        totalPriceEur:
          type: integer
          example: 100
        clientPriceInfo:
          type: integer
          example: 100
        totalPriceCalcInfo:
          type: integer
          example: 100
        agency:
          $ref: '#/components/schemas/AgencyDto'
        location:
          $ref: '#/components/schemas/LocationDto'
        yachtImages:
          type: array
          items:
            $ref: '#/components/schemas/YachtImageDto'
        sysDescription:
          type: string
          example: very good yacht
        fixedPrice:
          type: integer
          example: 100
        locations:
          type: array
          items:
            $ref: '#/components/schemas/LocationViewDto'
    YachtSearchResponseDto:
      type: object
      required:
        - name
      properties:
        id:
          type: integer
          format: int64
          example: 10
        name:
          type: string
          example: Big Blue Yacht
        location:
          $ref: '#/components/schemas/LocationDto'
        charterType:
          type: string
        vesselType:
          type: integer
          $ref: '#/components/schemas/VesselType'
        buildYear:
          type: integer
          example: 2011
        maxPersons:
          type: integer
          example: 6
        cabins:
          type: integer
          example: 4
        length:
          type: integer
          example: 15
        clientPriceEur:
          type: integer
          example: 100
        clientPriceInfo:
          type: integer
          example: 100
...
    AgencyDto:
      type: object
      required:
        - name
        - active
        - skipExternalSystem
      properties:
        id:
          type: integer
          format: int64
          example: 10
        name:
          type: string
          example: Najbolja agencija
        address:
          type: string
          example: Domovinskog rata 5
        city:
          type: string
          example: Split
        country:
          type: string
          example: Croatia
        zip:
          type: string
          example: 21000
        vatCode:
          type: string
          example: B59915249
        web:
          type: string
          example: www.sealand.it
        email:
          type: string
          example: booking@caymangroupcharter.it
        phone:
          type: string
          example: +39 3936101443
        mobile:
          type: string
          example: +39 3936101443
        iban:
          type: string
          example: 12347890
        active:
          type: string
          example: 12347890
        discount:
          type: integer
          example: 10
        director:
          type: string
          example: Konstantina Chachali
        skipExternalSystem:
          type: boolean
          example: true
    LocationDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        countryCode:
          type: string
          example: HR
        lat:
          type: integer
          example: 55
        lon:
          type: integer
          example: 55
    LocationViewDto:
      type: object
      properties:
        id:
          type: string
          example: c-55
        realId:
          type: integer
          format: int64
          example: 55
        name:
          type: string
          example: Croatia
        locationType:
          type: string
          example: COUNTRY
        countryCode:
          type: string
          example: HR
    OfferDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        dateFrom:
          type: string
          format: date
          example: 2025-04-25
        dateTo:
          type: string
          format: date
          example: 2025-04-25
        clientPriceEur:
          type: integer
          example: 55
        totalPriceEur:
          type: integer
          example: 55
        clientPriceInfo:
          type: integer
          example: 55
        totalPriceInfo:
          type: integer
          example: 55
        status:
          type: integer
          $ref: '#/components/schemas/OfferStatus'
        isFocused:
          type: boolean
          example: true
    YachtImageDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        url:
          type: string
          example: https://www.booking-manager.com/cbm/documents/6259149640000107392_lagoon-50--3.jpg
        position:
          type: integer
          example: 1
        mainImage:
          type: boolean
          example: true
...
security:
  - bearerAuth: []